+++ compass/src/components/Application/ApplicationDetails/ApplicationDetailsScenarios/ApplicationScenarioModal.container.js
@@ -1,13 +1,9 @@
 import { compose, graphql } from 'react-apollo';
-import AssignScenarioModal from './../../../Shared/AssignScenarioModal/AssignScenarioModal.component';
-import { GET_SCENARIOS } from './../../../Shared/gql';
+import AssignScenarioModal from '../../../Shared/AssignScenario/AssignScenarioModal.component';
 import { SET_APPLICATION_SCENARIOS } from './../../gql';
 import { SEND_NOTIFICATION } from '../../../../gql';
 
 export default compose(
-  graphql(GET_SCENARIOS, {
-    name: 'scenariosQuery',
-  }),
   graphql(SET_APPLICATION_SCENARIOS, {
     props: props => ({
       updateScenarios: async (applicationId, scenarios) => {
