+++ compass/src/components/Application/ApplicationDetails/CreateAPIModal/Forms/CredentialsForm.js
@@ -27,21 +27,18 @@ export default function CredentialsForm({ updateState }) {
     <form>
       <TextFormItem
         inputKey="client-id"
-        required
         type="password"
         label="Client ID"
         onChange={e => updateOAuth('clientId', e.target.value)}
       />
       <TextFormItem
         inputKey="client-secret"
-        required
         type="password"
         label="Client Secret"
         onChange={e => updateOAuth('clientSecret', e.target.value)}
       />
       <TextFormItem
         inputKey="url"
-        required
         type="url"
         label="Url"
         onChange={e => updateOAuth('url', e.target.value)}
