+++ compass/src/components/Shared/AssignScenario/AssignScenarioForm.container.js
@@ -1,0 +1,9 @@
+import { compose, graphql } from 'react-apollo';
+import { GET_SCENARIOS } from './../gql';
+import AssignScenarioForm from './AssignScenarioForm.component';
+
+export default compose(
+  graphql(GET_SCENARIOS, {
+    name: 'availableScenariosQuery',
+  }),
+)(AssignScenarioForm);
