+++ compass/src/components/Shared/AssignScenario/style.scss
